+++ cmd/jetmon2/main.go
@@ -71,6 +71,8 @@ func main() {
 		cmdAPI(os.Args[2:])
 	case "site-tenants":
 		cmdSiteTenants(os.Args[2:])
+	case "telemetry":
+		cmdTelemetry(os.Args[2:])
 	case "rollout":
 		cmdRollout(os.Args[2:])
 	default: